feat: Refactor user management by consolidating command and query services into UserService
